pyproject.toml
[project]
name = "pynite-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Connor Ferster", email = "connorferster@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.1",
    "plotly>=6.2.0",
    "plotly-3d-primitives>=0.4.0",
    "pydantic>=2.11.7",
    "pynitefea>=1.1.2",
    "numpydantic>=1.6.9",
    "deepmerge>=2.0",
]

[project.scripts]
pynite-tools = "pynite_tools:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
